refactor(feed-sort): move view model to 'option_selection' package
